feat: refactor branding logic into Branding component and remove useBranding hook
